+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Sort backups from newest to oldest in manage backups
 - Android: fix display of external links from Bitrefill
 - fix language sometimes not persistent across app restarts
+- Android: make the UI work with responsive font sizes and adhere to OS font size settings 
 
 ## v4.49.0
 - Bundle BitBox02 Nova firmware version v9.24.0